pyproject.toml
[build-system]
requires = ["setuptools >= 64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pddl_planner"
version = "0.1"
dependencies = [
  "pddl==0.4.0",
  "openai",
  "pyyaml",
  "opencv-python",
  "pyDatalog",
  "google-genai",
  "python-dotenv"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pddl_planner", "pddl_planner.*"]
exclude = ["img*", "files*"]